requirements.txt
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.18.0
streamlit==1.22.0
pillow==9.5.0
scikit-learn==1.2.2
koreanize-matplotlib==0.1.1 